requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
anthropic==0.18.1
python-dotenv==1.0.0
Pillow==10.2.0
requests==2.31.0
